+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on and uses the types of changes according to [Keep a Change
 - Updated `Invoke-M365SecurityAudit` to make `TenantAdminUrl` an optional parameter.
 - Improved connection handling and error messaging in `Connect-M365Suite`.
 - Enhanced `Invoke-M365SecurityAudit` to allow flexible inclusion and exclusion of specific recommendations, IG filters, and profile levels.
+- SupportsShoudProcess to also bypass connection checks in `Invoke-M365SecurityAudit` as well as Disconnect-M365Suite.
 
 ## [0.1.5] - 2024-06-08